refactor: move no-code elements in the code directory
