+++ ui/src/components/code/components/tasks/ClearButton.vue
