+++ ui/src/components/code/components/tasks/TaskComplex.vue
@@ -17,7 +17,7 @@
 </script>
 
 <script>
-    import Task from "./Task";
+    import Task from "./MixinTask";
 
     export default {
         inheritAttrs: false,
