+++ ui/src/components/code/components/tasks/TaskCondition.vue
@@ -41,7 +41,7 @@
 </script>
 
 <script>
-    import Task from "./Task"
+    import Task from "./MixinTask"
 
     export default {
         mixins: [Task],
