+++ ui/src/components/code/components/tasks/TaskExpression.vue
@@ -11,7 +11,7 @@
     />
 </template>
 <script lang="ts" setup>
-    import {collapseEmptyValues} from "./Task";
+    import {collapseEmptyValues} from "./MixinTask";
     import Editor from "../../../components/inputs/Editor.vue";
     import * as YAML_UTILS from "@kestra-io/ui-libs/flow-yaml-utils";
     import {computed, ref} from "vue";
