+++ ui/src/components/code/components/tasks/TaskNumber.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-    import Task from "./Task"
+    import Task from "./MixinTask"
     export default {
         mixins: [Task],
         computed: {
