+++ ui/src/components/code/components/tasks/TaskSubflowId.vue
@@ -20,7 +20,7 @@
 <script>
     import {mapStores} from "pinia";
     import {useFlowStore} from "../../../stores/flow";
-    import Task from "./Task";
+    import Task from "./MixinTask";
 
     export default {
         mixins: [Task],
